refactor: refactored types.ts to seperate files
