+++ src/api/workout/useGetAllWorkouts.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { Workout } from "../../types/types";
+import { Workout } from "../../types/MuscleInfo";
 import axios from "axios";
 
 export const useGetAllExercises = () => {
