+++ src/types/MuscleInfo.ts
@@ -1,0 +1,4 @@
+export type MuscleInfo = {
+  id: string;
+  name: string;
+};
