+++ src/types/WorkoutExercise.ts
@@ -1,0 +1,7 @@
+import { Exercise } from "./Exercise";
+
+export interface WorkoutExercise {
+  exercise: Exercise;
+  sets: number;
+  repsPerSet: number;
+}
